support, mapper: add llvm-support, more boilerplate
